+++ packages/tldraw/src/lib/shapes/geo/GeoShapeUtil.tsx
@@ -15,7 +15,6 @@ import {
 	SVGContainer,
 	Stadium2d,
 	SvgExportContext,
-	TLDefaultDashStyle,
 	TLGeoShape,
 	TLOnEditEndHandler,
 	TLOnResizeHandler,
@@ -31,6 +30,7 @@ import {
 
 import { HyperlinkButton } from '../shared/HyperlinkButton'
 import { TextLabel } from '../shared/TextLabel'
+import { createTextSvgStringFromSpans } from '../shared/createTextSvgStringFromSpans'
 import {
 	FONT_FAMILIES,
 	LABEL_FONT_SIZES,
@@ -42,24 +42,22 @@ import {
 	getFillDefForExport,
 	getFontDefForExport,
 } from '../shared/defaultStyleDefs'
-import { getTextLabelSvgElement } from '../shared/getTextLabelSvgElement'
 import { getRoundedInkyPolygonPath, getRoundedPolygonPoints } from '../shared/polygon-helpers'
-import { cloudOutline, cloudSvgPath } from './cloudOutline'
-import { DashStyleCloud, DashStyleCloudSvg } from './components/DashStyleCloud'
-import { DashStyleEllipse, DashStyleEllipseSvg } from './components/DashStyleEllipse'
-import { DashStyleOval, DashStyleOvalSvg } from './components/DashStyleOval'
-import { DashStylePolygon, DashStylePolygonSvg } from './components/DashStylePolygon'
-import { DrawStyleCloud, DrawStyleCloudSvg } from './components/DrawStyleCloud'
-import { DrawStyleEllipseSvg, getEllipseIndicatorPath } from './components/DrawStyleEllipse'
-import { DrawStylePolygon, DrawStylePolygonSvg } from './components/DrawStylePolygon'
-import { SolidStyleCloud, SolidStyleCloudSvg } from './components/SolidStyleCloud'
-import { SolidStyleEllipse, SolidStyleEllipseSvg } from './components/SolidStyleEllipse'
-import {
-	SolidStyleOval,
-	SolidStyleOvalSvg,
-	getOvalIndicatorPath,
-} from './components/SolidStyleOval'
-import { SolidStylePolygon, SolidStylePolygonSvg } from './components/SolidStylePolygon'
+import { getSvgFromString } from '../shared/svgs'
+import { cloudOutline, cloudSvgPath } from './cloud-helpers'
+import { DashStyleCloud } from './components/DashStyleCloud'
+import { DashStyleEllipse } from './components/DashStyleEllipse'
+import { DashStyleOval } from './components/DashStyleOval'
+import { DashStylePolygon } from './components/DashStylePolygon'
+import { DrawStyleCloud } from './components/DrawStyleCloud'
+import { getEllipseIndicatorPath } from './components/DrawStyleEllipse'
+import { DrawStylePolygon } from './components/DrawStylePolygon'
+import { SolidStyleCloud } from './components/SolidStyleCloud'
+import { SolidStyleEllipse } from './components/SolidStyleEllipse'
+import { SolidStyleOval, getOvalIndicatorPath } from './components/SolidStyleOval'
+import { SolidStylePolygon } from './components/SolidStylePolygon'
+import { getCloudSvg, getEllipseSvg, getOvalSvg, getPolygonSvg } from './geo-svgs'
+import { getLines } from './lines'
 
 const LABEL_PADDING = 16
 const MIN_SIZE_WITH_LABEL = 17 * 3
@@ -449,11 +447,7 @@ export class GeoShapeUtil extends BaseBoxShapeUtil<TLGeoShape> {
 					break
 				}
 				case 'ellipse': {
-					if (dash === 'solid') {
-						return (
-							<SolidStyleEllipse strokeWidth={strokeWidth} w={w} h={h} color={color} fill={fill} />
-						)
-					} else if (dash === 'dashed' || dash === 'dotted') {
+					if (dash === 'dashed' || dash === 'dotted') {
 						return (
 							<DashStyleEllipse
 								id={id}
@@ -465,12 +459,11 @@ export class GeoShapeUtil extends BaseBoxShapeUtil<TLGeoShape> {
 								fill={fill}
 							/>
 						)
-					} else if (dash === 'draw') {
+					} else {
 						return (
 							<SolidStyleEllipse strokeWidth={strokeWidth} w={w} h={h} color={color} fill={fill} />
 						)
 					}
-					break
 				}
 				case 'oval': {
 					if (dash === 'solid') {
@@ -616,222 +609,74 @@ export class GeoShapeUtil extends BaseBoxShapeUtil<TLGeoShape> {
 	}
 
 	override toSvg(shape: TLGeoShape, ctx: SvgExportContext) {
-		const { id, props } = shape
-		const strokeWidth = STROKE_SIZES[props.size]
 		const theme = getDefaultColorTheme({ isDarkMode: ctx.isDarkMode })
 		ctx.addExportDef(getFillDefForExport(shape.props.fill, theme))
 
-		let svgElm: SVGElement
+		let labelString = ''
+		let shapeString = ''
 
-		switch (props.geo) {
+		switch (shape.props.geo) {
 			case 'ellipse': {
-				switch (props.dash) {
-					case 'draw':
-						svgElm = DrawStyleEllipseSvg({
-							id,
-							w: props.w,
-							h: props.h,
-							color: props.color,
-							fill: props.fill,
-							strokeWidth,
-							theme,
-						})
-						break
-
-					case 'solid':
-						svgElm = SolidStyleEllipseSvg({
-							strokeWidth,
-							w: props.w,
-							h: props.h,
-							color: props.color,
-							fill: props.fill,
-							theme,
-						})
-						break
-
-					default:
-						svgElm = DashStyleEllipseSvg({
-							id,
-							strokeWidth,
-							w: props.w,
-							h: props.h,
-							dash: props.dash,
-							color: props.color,
-							fill: props.fill,
-							theme,
-						})
-						break
-				}
+				shapeString = getEllipseSvg(shape, theme)
 				break
 			}
-
 			case 'oval': {
-				switch (props.dash) {
-					case 'draw':
-						svgElm = DashStyleOvalSvg({
-							id,
-							strokeWidth,
-							w: props.w,
-							h: props.h,
-							dash: props.dash,
-							color: props.color,
-							fill: props.fill,
-							theme,
-						})
-						break
-
-					case 'solid':
-						svgElm = SolidStyleOvalSvg({
-							strokeWidth,
-							w: props.w,
-							h: props.h,
-							color: props.color,
-							fill: props.fill,
-							theme,
-						})
-						break
-
-					default:
-						svgElm = DashStyleOvalSvg({
-							id,
-							strokeWidth,
-							w: props.w,
-							h: props.h,
-							dash: props.dash,
-							color: props.color,
-							fill: props.fill,
-							theme,
-						})
-				}
+				shapeString = getOvalSvg(shape, theme)
 				break
 			}
-
 			case 'cloud': {
-				switch (props.dash) {
-					case 'draw':
-						svgElm = DrawStyleCloudSvg({
-							id,
-							strokeWidth,
-							w: props.w,
-							h: props.h,
-							color: props.color,
-							fill: props.fill,
-							size: props.size,
-							theme,
-						})
-						break
-
-					case 'solid':
-						svgElm = SolidStyleCloudSvg({
-							strokeWidth,
-							w: props.w,
-							h: props.h,
-							color: props.color,
-							fill: props.fill,
-							size: props.size,
-							id,
-							theme,
-						})
-						break
-
-					default:
-						svgElm = DashStyleCloudSvg({
-							id,
-							strokeWidth,
-							w: props.w,
-							h: props.h,
-							dash: props.dash,
-							color: props.color,
-							fill: props.fill,
-							theme,
-							size: props.size,
-						})
-				}
+				shapeString = getCloudSvg(shape, theme)
 				break
 			}
 			default: {
-				const geometry = this.editor.getShapeGeometry(shape)
-				const outline =
-					geometry instanceof Group2d ? geometry.children[0].vertices : geometry.vertices
-				const lines = getLines(shape.props, strokeWidth)
-
-				switch (props.dash) {
-					case 'draw':
-						svgElm = DrawStylePolygonSvg({
-							id,
-							fill: props.fill,
-							color: props.color,
-							strokeWidth,
-							outline,
-							lines,
-							theme,
-						})
-						break
-
-					case 'solid':
-						svgElm = SolidStylePolygonSvg({
-							fill: props.fill,
-							color: props.color,
-							strokeWidth,
-							outline,
-							lines,
-							theme,
-						})
-						break
-
-					default:
-						svgElm = DashStylePolygonSvg({
-							dash: props.dash,
-							fill: props.fill,
-							color: props.color,
-							strokeWidth,
-							outline,
-							lines,
-							theme,
-						})
-						break
-				}
+				shapeString = getPolygonSvg(this.editor, shape, theme)
 				break
 			}
 		}
 
-		if (props.text) {
+		if (shape.props.text) {
 			const bounds = this.editor.getShapeGeometry(shape).bounds
 
 			ctx.addExportDef(getFontDefForExport(shape.props.font))
 
-			const rootTextElm = getTextLabelSvgElement({
-				editor: this.editor,
-				shape,
-				font: DefaultFontFamilies[shape.props.font],
-				bounds,
+			const padding = 16
+
+			const opts = {
+				fontSize: LABEL_FONT_SIZES[shape.props.size],
+				fontFamily: DefaultFontFamilies[shape.props.font],
+				textAlign: shape.props.align,
+				verticalTextAlign: shape.props.verticalAlign,
+				width: Math.ceil(bounds.width),
+				height: Math.ceil(bounds.height),
+				padding,
+				lineHeight: TEXT_PROPS.lineHeight,
+				fontStyle: 'normal',
+				fontWeight: 'normal',
+				overflow: 'wrap' as const,
+				offsetX: 0,
+			}
+
+			const spans = this.editor.textMeasure.measureTextSpans(shape.props.text, opts)
+			const textElm = createTextSvgStringFromSpans(spans, {
+				...opts,
+				fill: theme[shape.props.labelColor].solid,
+			})
+			const textBgElm = createTextSvgStringFromSpans(spans, {
+				...opts,
+				fill: theme.background,
+				stroke: theme.background,
+				strokeWidth: 2,
 			})
 
-			const textElm = rootTextElm.cloneNode(true) as SVGTextElement
-			textElm.setAttribute('fill', theme[shape.props.labelColor].solid)
-			textElm.setAttribute('stroke', 'none')
-
-			const textBgEl = rootTextElm.cloneNode(true) as SVGTextElement
-			textBgEl.setAttribute('stroke-width', '2')
-			textBgEl.setAttribute('fill', theme.background)
-			textBgEl.setAttribute('stroke', theme.background)
-
-			const groupEl = document.createElementNS('http://www.w3.org/2000/svg', 'g')
-			groupEl.append(textBgEl)
-			groupEl.append(textElm)
-
-			if (svgElm.nodeName === 'g') {
-				svgElm.appendChild(groupEl)
-				return svgElm
-			} else {
-				const g = document.createElementNS('http://www.w3.org/2000/svg', 'g')
-				g.appendChild(svgElm)
-				g.appendChild(groupEl)
-				return g
-			}
+			labelString = `<g>${textBgElm}${textElm}</g>`
 		}
 
-		return svgElm
+		const result =
+			labelString && shapeString
+				? `<g>${shapeString}${labelString}</g>`
+				: shapeString || labelString
+
+		return getSvgFromString(result)
 	}
 
 	override getCanvasSvgDefs(): TLShapeUtilCanvasSvgDef[] {
@@ -1109,67 +954,6 @@ function getLabelSize(editor: Editor, shape: TLGeoShape) {
 		w: size.w + LABEL_PADDING * 2,
 		h: size.h + LABEL_PADDING * 2,
 	}
-}
-
-function getLines(props: TLGeoShape['props'], sw: number) {
-	switch (props.geo) {
-		case 'x-box': {
-			return getXBoxLines(props.w, props.h, sw, props.dash)
-		}
-		case 'check-box': {
-			return getCheckBoxLines(props.w, props.h)
-		}
-		default: {
-			return undefined
-		}
-	}
-}
-
-function getXBoxLines(w: number, h: number, sw: number, dash: TLDefaultDashStyle) {
-	const inset = dash === 'draw' ? 0.62 : 0
-
-	if (dash === 'dashed') {
-		return [
-			[new Vec(0, 0), new Vec(w / 2, h / 2)],
-			[new Vec(w, h), new Vec(w / 2, h / 2)],
-			[new Vec(0, h), new Vec(w / 2, h / 2)],
-			[new Vec(w, 0), new Vec(w / 2, h / 2)],
-		]
-	}
-
-	const clampX = (x: number) => Math.max(0, Math.min(w, x))
-	const clampY = (y: number) => Math.max(0, Math.min(h, y))
-
-	return [
-		[
-			new Vec(clampX(sw * inset), clampY(sw * inset)),
-			new Vec(clampX(w - sw * inset), clampY(h - sw * inset)),
-		],
-		[
-			new Vec(clampX(sw * inset), clampY(h - sw * inset)),
-			new Vec(clampX(w - sw * inset), clampY(sw * inset)),
-		],
-	]
-}
-
-function getCheckBoxLines(w: number, h: number) {
-	const size = Math.min(w, h) * 0.82
-	const ox = (w - size) / 2
-	const oy = (h - size) / 2
-
-	const clampX = (x: number) => Math.max(0, Math.min(w, x))
-	const clampY = (y: number) => Math.max(0, Math.min(h, y))
-
-	return [
-		[
-			new Vec(clampX(ox + size * 0.25), clampY(oy + size * 0.52)),
-			new Vec(clampX(ox + size * 0.45), clampY(oy + size * 0.82)),
-		],
-		[
-			new Vec(clampX(ox + size * 0.45), clampY(oy + size * 0.82)),
-			new Vec(clampX(ox + size * 0.82), clampY(oy + size * 0.22)),
-		],
-	]
 }
 
 /**
