+++ packages/tldraw/src/lib/shapes/geo/components/DrawStyleCloud.tsx
@@ -1,12 +1,7 @@
-import { TLDefaultColorTheme, TLGeoShape, TLShapeId } from '@tldraw/editor'
+import { TLGeoShape, TLShapeId } from '@tldraw/editor'
 import * as React from 'react'
-import {
-	ShapeFill,
-	getShapeFillSvg,
-	getSvgWithShapeFill,
-	useDefaultColorTheme,
-} from '../../shared/ShapeFill'
-import { inkyCloudSvgPath } from '../cloudOutline'
+import { ShapeFill, useDefaultColorTheme } from '../../shared/ShapeFill'
+import { inkyCloudSvgPath } from '../cloud-helpers'
 
 export const DrawStyleCloud = React.memo(function StyleCloud({
 	fill,
@@ -30,36 +25,3 @@ export const DrawStyleCloud = React.memo(function StyleCloud({
 		</>
 	)
 })
-
-export function DrawStyleCloudSvg({
-	fill,
-	color,
-	strokeWidth,
-	theme,
-	w,
-	h,
-	id,
-	size,
-}: Pick<TLGeoShape['props'], 'fill' | 'color' | 'w' | 'h' | 'size'> & {
-	strokeWidth: number
-	theme: TLDefaultColorTheme
-	id: TLShapeId
-}) {
-	const pathData = inkyCloudSvgPath(w, h, id, size)
-
-	const strokeElement = document.createElementNS('http://www.w3.org/2000/svg', 'path')
-	strokeElement.setAttribute('d', pathData)
-	strokeElement.setAttribute('stroke-width', strokeWidth.toString())
-	strokeElement.setAttribute('stroke', theme[color].solid)
-	strokeElement.setAttribute('fill', 'none')
-
-	// Get the fill element, if any
-	const fillElement = getShapeFillSvg({
-		d: pathData,
-		fill,
-		color,
-		theme,
-	})
-
-	return getSvgWithShapeFill(strokeElement, fillElement)
-}
