+++ packages/tldraw/src/lib/shapes/geo/components/SolidStyleCloud.tsx
@@ -1,12 +1,7 @@
-import { TLDefaultColorTheme, TLGeoShape, TLShapeId } from '@tldraw/editor'
+import { TLGeoShape, TLShapeId } from '@tldraw/editor'
 import * as React from 'react'
-import {
-	ShapeFill,
-	getShapeFillSvg,
-	getSvgWithShapeFill,
-	useDefaultColorTheme,
-} from '../../shared/ShapeFill'
-import { cloudSvgPath } from '../cloudOutline'
+import { ShapeFill, useDefaultColorTheme } from '../../shared/ShapeFill'
+import { cloudSvgPath } from '../cloud-helpers'
 
 export const SolidStyleCloud = React.memo(function SolidStyleCloud({
 	fill,
@@ -30,36 +25,3 @@ export const SolidStyleCloud = React.memo(function SolidStyleCloud({
 		</>
 	)
 })
-
-export function SolidStyleCloudSvg({
-	fill,
-	color,
-	strokeWidth,
-	theme,
-	w,
-	h,
-	id,
-	size,
-}: Pick<TLGeoShape['props'], 'fill' | 'color' | 'w' | 'h' | 'size'> & {
-	strokeWidth: number
-	theme: TLDefaultColorTheme
-	id: TLShapeId
-}) {
-	const pathData = cloudSvgPath(w, h, id, size)
-
-	const strokeElement = document.createElementNS('http://www.w3.org/2000/svg', 'path')
-	strokeElement.setAttribute('d', pathData)
-	strokeElement.setAttribute('stroke-width', strokeWidth.toString())
-	strokeElement.setAttribute('stroke', theme[color].solid)
-	strokeElement.setAttribute('fill', 'none')
-
-	// Get the fill element, if any
-	const fillElement = getShapeFillSvg({
-		d: pathData,
-		fill,
-		color,
-		theme,
-	})
-
-	return getSvgWithShapeFill(strokeElement, fillElement)
-}
