+++ packages/tldraw/src/lib/shapes/geo/oval-helpers.ts
