+++ packages/tldraw/src/lib/shapes/shared/ShapeFill.tsx
@@ -60,54 +60,18 @@ const PatternFill = function PatternFill({ d, color, theme }: ShapeFillProps) {
 	)
 }
 
-export function getShapeFillSvg({ d, color, fill, theme }: ShapeFillProps) {
-	if (fill === 'none') {
-		return
-	}
-
-	if (fill === 'pattern') {
-		const gEl = document.createElementNS('http://www.w3.org/2000/svg', 'g')
-		const path1El = document.createElementNS('http://www.w3.org/2000/svg', 'path')
-		path1El.setAttribute('d', d)
-		path1El.setAttribute('fill', theme[color].pattern)
-
-		const path2El = document.createElementNS('http://www.w3.org/2000/svg', 'path')
-		path2El.setAttribute('d', d)
-		path2El.setAttribute('fill', `url(#hash_pattern)`)
-
-		gEl.appendChild(path1El)
-		gEl.appendChild(path2El)
-		return gEl
-	}
-
-	const path = document.createElementNS('http://www.w3.org/2000/svg', 'path')
-	path.setAttribute('d', d)
-
+export function getShapeFillSvgString({ d, color, fill, theme }: ShapeFillProps) {
 	switch (fill) {
-		case 'semi': {
-			path.setAttribute('fill', theme.solid)
-			break
-		}
-		case 'solid': {
-			{
-				path.setAttribute('fill', theme[color].semi)
-			}
-			break
-		}
+		case 'pattern':
+			return `<g><path fill="${theme[color].pattern}" d="${d}"/><path fill="url(#hash_pattern)" d="${d}"/></g>`
+		case 'semi':
+			return `<path d="${d}" fill="${theme.solid}"/>`
+		case 'solid':
+			return `<path d="${d}" fill="${theme[color].semi}"/>`
 	}
-
-	return path
+	return ''
 }
 
-export function getSvgWithShapeFill(foregroundPath: SVGElement, backgroundPath?: SVGElement) {
-	if (backgroundPath) {
-		// If there is a fill element, return a group containing the fill and the path
-		const g = document.createElementNS('http://www.w3.org/2000/svg', 'g')
-		g.appendChild(backgroundPath)
-		g.appendChild(foregroundPath)
-		return g
-	} else {
-		// Otherwise, just return the path
-		return foregroundPath
-	}
+export function getSvgStringWithShapeFill(foregroundPath: string, backgroundPath?: string) {
+	return backgroundPath ? `<g>${backgroundPath}${foregroundPath}</g>` : foregroundPath
 }
